requirements.txt
django
pandas
numpy
sklearn.metrics
statsmodels
matplotlib
django-import-export
PyMySQL
sqlalchemy
django-crontab
django-livesync